:herb: Do npm install in command.
Previously the npm script watch-dev and watch test were failing with a
nodemon not found and mocha not found error. This is probably to do
with the path as npm add $(npm bin) to the path to allow node modules
to be run directly.

This seems to resolve it but there may be better ways of resolving this.